+++ adapters/handlers/rest/authz/handlers_authz_add_permission_test.go
@@ -9,13 +9,12 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rbacrest
+package authz
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -28,11 +27,11 @@ import (
 	schemaMocks "github.com/weaviate/weaviate/usecases/schema/mocks"
 )
 
-func TestCreateRoleSuccess(t *testing.T) {
+func TestAddPermissionsSuccess(t *testing.T) {
 	type testCase struct {
 		name      string
 		principal *models.Principal
-		params    authz.CreateRoleParams
+		params    authz.AddPermissionsParams
 		// readCollection              bool
 		// readTenant                  bool
 		// readTenantWithoutCollection bool
@@ -42,9 +41,9 @@ func TestCreateRoleSuccess(t *testing.T) {
 		{
 			name:      "all are *",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
-					Name: String("newRole"),
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
+					Name: String("test"),
 					Permissions: []*models.Permission{
 						{
 							Action: String(authorization.CreateSchema),
@@ -56,8 +55,8 @@ func TestCreateRoleSuccess(t *testing.T) {
 		{
 			name:      "collection checks",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
 					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
@@ -72,8 +71,8 @@ func TestCreateRoleSuccess(t *testing.T) {
 		{
 			name:      "collection and tenant checks",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
 					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
@@ -90,8 +89,8 @@ func TestCreateRoleSuccess(t *testing.T) {
 		{
 			name:      "* collections and tenant checks",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
 					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
@@ -112,9 +111,14 @@ func TestCreateRoleSuccess(t *testing.T) {
 			schemaReader := schemaMocks.NewSchemaGetter(t)
 			logger, _ := test.NewNullLogger()
 
-			authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
-			controller.On("GetRoles", *tt.params.Body.Name).Return(map[string][]authorization.Policy{}, nil)
-			controller.On("UpsertRolesPermissions", mock.Anything).Return(nil)
+			policies, err := conv.RolesToPolicies(&models.Role{
+				Name:        tt.params.Body.Name,
+				Permissions: tt.params.Body.Permissions,
+			})
+			require.Nil(t, err)
+
+			authorizer.On("Authorize", tt.principal, authorization.UPDATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(nil)
+			controller.On("UpsertRolesPermissions", policies).Return(nil)
 
 			// if tt.readCollection {
 			// 	schemaReader.On("ReadOnlyClass",
@@ -146,49 +150,19 @@ func TestCreateRoleSuccess(t *testing.T) {
 				schemaReader: schemaReader,
 				logger:       logger,
 			}
-			res := h.createRole(tt.params, tt.principal)
-			parsed, ok := res.(*authz.CreateRoleCreated)
+			res := h.addPermissions(tt.params, tt.principal)
+			parsed, ok := res.(*authz.AddPermissionsOK)
 			assert.True(t, ok)
 			assert.NotNil(t, parsed)
 		})
 	}
 }
 
-func TestCreateRoleConflict(t *testing.T) {
-	authorizer := mocks.NewAuthorizer(t)
-	controller := mocks.NewController(t)
-	logger, _ := test.NewNullLogger()
-
-	principal := &models.Principal{Username: "user1"}
-	params := authz.CreateRoleParams{
-		Body: &models.Role{
-			Name: String("newRole"),
-			Permissions: []*models.Permission{
-				{
-					Action: String(authorization.CreateSchema),
-				},
-			},
-		},
-	}
-	authorizer.On("Authorize", principal, authorization.CREATE, authorization.Roles("newRole")[0]).Return(nil)
-	controller.On("GetRoles", *params.Body.Name).Return(map[string][]authorization.Policy{"newRole": {}}, nil)
-
-	h := &authZHandlers{
-		authorizer: authorizer,
-		controller: controller,
-		logger:     logger,
-	}
-	res := h.createRole(params, principal)
-	parsed, ok := res.(*authz.CreateRoleConflict)
-	assert.True(t, ok)
-	assert.Contains(t, parsed.Payload.Error[0].Message, fmt.Sprintf("role with name %s already exists", *params.Body.Name))
-}
-
-func TestCreateRoleBadRequest(t *testing.T) {
+func TestAddPermissionsBadRequest(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.CreateRoleParams
-		upsertErr     error
+		params        authz.AddPermissionsParams
+		principal     *models.Principal
 		expectedError string
 		// readCollection              bool
 		// readTenant                  bool
@@ -198,8 +172,8 @@ func TestCreateRoleBadRequest(t *testing.T) {
 	tests := []testCase{
 		{
 			name: "role name is required",
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
 					Name: String(""),
 					Permissions: []*models.Permission{
 						{
@@ -208,36 +182,39 @@ func TestCreateRoleBadRequest(t *testing.T) {
 					},
 				},
 			},
+			principal:     &models.Principal{Username: "user1"},
 			expectedError: "role name is required",
 		},
 		{
 			name: "role has to have at least 1 permission",
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
-					Name:        String("newRole"),
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
+					Name:        String("someName"),
 					Permissions: []*models.Permission{},
 				},
 			},
+			principal:     &models.Principal{Username: "user1"},
 			expectedError: "role has to have at least 1 permission",
 		},
 		{
-			name: "invalid permission",
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
-					Name: String("someRole"),
+			name: "invalid action",
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
+					Name: String("someName"),
 					Permissions: []*models.Permission{
 						{
-							Action: String("manage_something"),
+							Action: String("manage_somethingelse"),
 						},
 					},
 				},
 			},
+			principal:     &models.Principal{Username: "user1"},
 			expectedError: "invalid permission",
 		},
 		{
-			name: "cannot create role with the same name as builtin role",
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
+			name: "update builtin role",
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
 					Name: &authorization.BuiltInRoles[0],
 					Permissions: []*models.Permission{
 						{
@@ -246,16 +223,17 @@ func TestCreateRoleBadRequest(t *testing.T) {
 					},
 				},
 			},
-			expectedError: "you can not create role with the same name as builtin role",
+			principal:     &models.Principal{Username: "user1"},
+			expectedError: "you can not update builtin role",
 		},
 		// {
 		// 	name: "collection doesn't exist",
-		// 	params: authz.CreateRoleParams{
-		// 		Body: &models.Role{
+		// 	params: authz.AddPermissionsParams{
+		// 		Body: authz.AddPermissionsBody{
 		// 			Name: String("newRole"),
 		// 			Permissions: []*models.Permission{
 		// 				{
-		// 					Action:     String(authorization.CreateSchema),
+		// 					Action:     String("manage_roles"),
 		// 					Collection: String("ABC"),
 		// 				},
 		// 			},
@@ -266,12 +244,12 @@ func TestCreateRoleBadRequest(t *testing.T) {
 		// },
 		// {
 		// 	name: "tenant doesn't exist",
-		// 	params: authz.CreateRoleParams{
-		// 		Body: &models.Role{
+		// 	params: authz.AddPermissionsParams{
+		// 		Body: authz.AddPermissionsBody{
 		// 			Name: String("newRole"),
 		// 			Permissions: []*models.Permission{
 		// 				{
-		// 					Action:     String(authorization.CreateSchema),
+		// 					Action:     String("manage_roles"),
 		// 					Collection: String("ABC"),
 		// 					Tenant:     String("Tenant1"),
 		// 				},
@@ -283,8 +261,8 @@ func TestCreateRoleBadRequest(t *testing.T) {
 		// },
 		// {
 		// 	name: "tenant doesn't exist with * collection",
-		// 	params: authz.CreateRoleParams{
-		// 		Body: &models.Role{
+		// 	params: authz.AddPermissionsParams{
+		// 		Body: authz.AddPermissionsBody{
 		// 			Name: String("newRole"),
 		// 			Permissions: []*models.Permission{
 		// 				{
@@ -301,16 +279,10 @@ func TestCreateRoleBadRequest(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			authorizer := mocks.NewAuthorizer(t)
 			controller := mocks.NewController(t)
+			authorizer := mocks.NewAuthorizer(t)
 			schemaReader := schemaMocks.NewSchemaGetter(t)
 			logger, _ := test.NewNullLogger()
-
-			if tt.expectedError == "" {
-				authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
-				controller.On("GetRoles", *tt.params.Body.Name).Return(map[string][]authorization.Policy{}, nil)
-				controller.On("UpsertRolesPermissions", mock.Anything).Return(tt.upsertErr)
-			}
 
 			// if tt.readCollection {
 			// 	authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
@@ -340,15 +312,14 @@ func TestCreateRoleBadRequest(t *testing.T) {
 			// 		Physical: map[string]sharding.Physical{},
 			// 	})
 			// }
-
 			h := &authZHandlers{
-				authorizer:   authorizer,
 				controller:   controller,
+				authorizer:   authorizer,
 				schemaReader: schemaReader,
 				logger:       logger,
 			}
-			res := h.createRole(tt.params, nil)
-			parsed, ok := res.(*authz.CreateRoleBadRequest)
+			res := h.addPermissions(tt.params, tt.principal)
+			parsed, ok := res.(*authz.AddPermissionsBadRequest)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -358,10 +329,10 @@ func TestCreateRoleBadRequest(t *testing.T) {
 	}
 }
 
-func TestCreateRoleForbidden(t *testing.T) {
+func TestAddPermissionsForbidden(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.CreateRoleParams
+		params        authz.AddPermissionsParams
 		principal     *models.Principal
 		authorizeErr  error
 		expectedError string
@@ -369,10 +340,10 @@ func TestCreateRoleForbidden(t *testing.T) {
 
 	tests := []testCase{
 		{
-			name: "authorization error",
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
-					Name: String("newRole"),
+			name: "update some role",
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
+					Name: String("someRole"),
 					Permissions: []*models.Permission{
 						{
 							Action: String(authorization.CreateSchema),
@@ -381,8 +352,8 @@ func TestCreateRoleForbidden(t *testing.T) {
 				},
 			},
 			principal:     &models.Principal{Username: "user1"},
-			authorizeErr:  errors.New("authorization error"),
-			expectedError: "authorization error",
+			authorizeErr:  fmt.Errorf("some error from authZ"),
+			expectedError: "some error from authZ",
 		},
 	}
 
@@ -392,15 +363,15 @@ func TestCreateRoleForbidden(t *testing.T) {
 			controller := mocks.NewController(t)
 			logger, _ := test.NewNullLogger()
 
-			authorizer.On("Authorize", tt.principal, authorization.CREATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(tt.authorizeErr)
+			authorizer.On("Authorize", tt.principal, authorization.UPDATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(tt.authorizeErr)
 
 			h := &authZHandlers{
 				authorizer: authorizer,
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.createRole(tt.params, tt.principal)
-			parsed, ok := res.(*authz.CreateRoleForbidden)
+			res := h.addPermissions(tt.params, tt.principal)
+			parsed, ok := res.(*authz.AddPermissionsForbidden)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -410,10 +381,10 @@ func TestCreateRoleForbidden(t *testing.T) {
 	}
 }
 
-func TestCreateRoleInternalServerError(t *testing.T) {
+func TestAddPermissionsInternalServerError(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.CreateRoleParams
+		params        authz.AddPermissionsParams
 		principal     *models.Principal
 		upsertErr     error
 		expectedError string
@@ -421,10 +392,10 @@ func TestCreateRoleInternalServerError(t *testing.T) {
 
 	tests := []testCase{
 		{
-			name: "upsert roles permissions error",
-			params: authz.CreateRoleParams{
-				Body: &models.Role{
-					Name: String("newRole"),
+			name: "update some role",
+			params: authz.AddPermissionsParams{
+				Body: authz.AddPermissionsBody{
+					Name: String("someRole"),
 					Permissions: []*models.Permission{
 						{
 							Action: String(authorization.CreateSchema),
@@ -433,8 +404,8 @@ func TestCreateRoleInternalServerError(t *testing.T) {
 				},
 			},
 			principal:     &models.Principal{Username: "user1"},
-			upsertErr:     errors.New("upsert error"),
-			expectedError: "upsert error",
+			upsertErr:     fmt.Errorf("some error from controller"),
+			expectedError: "some error from controller",
 		},
 	}
 
@@ -444,20 +415,16 @@ func TestCreateRoleInternalServerError(t *testing.T) {
 			controller := mocks.NewController(t)
 			logger, _ := test.NewNullLogger()
 
-			policies, err := conv.RolesToPolicies(tt.params.Body)
-			require.Nil(t, err)
-
-			authorizer.On("Authorize", tt.principal, authorization.CREATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(nil)
-			controller.On("GetRoles", *tt.params.Body.Name).Return(map[string][]authorization.Policy{}, nil)
-			controller.On("UpsertRolesPermissions", policies).Return(tt.upsertErr)
+			authorizer.On("Authorize", tt.principal, authorization.UPDATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(nil)
+			controller.On("UpsertRolesPermissions", mock.Anything).Return(tt.upsertErr)
 
 			h := &authZHandlers{
 				authorizer: authorizer,
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.createRole(tt.params, tt.principal)
-			parsed, ok := res.(*authz.CreateRoleInternalServerError)
+			res := h.addPermissions(tt.params, tt.principal)
+			parsed, ok := res.(*authz.AddPermissionsInternalServerError)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -465,8 +432,4 @@ func TestCreateRoleInternalServerError(t *testing.T) {
 			}
 		})
 	}
-}
-
-func String(s string) *string {
-	return &s
 }
