+++ adapters/handlers/rest/authz/handlers_authz_delete_role_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rbacrest
+package authz
 
 import (
 	"testing"
