+++ adapters/handlers/rest/authz/handlers_authz_get_roles_for_user_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rbacrest
+package authz
 
 import (
 	"fmt"
